Fix compiler warning about wrong printf specification on 64 bit systems.
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1531 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/base/idlist64.c
@@ -716,7 +716,7 @@ uint64_t GWEN_IdList64_GetIdAt(const GWEN_IDLIST64 *idl, uint64_t index) {
   } /* while */
 
   if (!idt) {
-    DBG_INFO(GWEN_LOGDOMAIN, "Index %lld not found", index);
+    DBG_INFO(GWEN_LOGDOMAIN, "Index %lld not found", (long long)index);
     return 0;
   }
